requirements.txt
robotframework == 6.0.2
robotframework-seleniumlibrary == 6.0.0
webdrivermanager == 0.10.0
python-dotenv == 0.21.0